device C化bugfix
# message auto-generated for no-merge-commit merge:
merge device_c into master

[AR20240408421922]device c化bugfix

Created-by: msprofDeveloper

Match-id-642b90e99dbb54b4a7632c0a1aa835a649b56c19

+++ analysis/csrc/domain/entities/hal/include/top_down_task.h
@@ -32,7 +32,7 @@ struct TopDownTask {
     TopDownTask() = default;
     TopDownTask(uint16_t taskId, uint16_t batchId, uint32_t streamId, uint32_t contextId, int32_t indexId,
                 std::string deviceType, std::string hostType, uint64_t modelId, int64_t connectionId,
-                uint64_t startTime, uint64_t endTime)
+                double startTime, double endTime)
         : taskId(taskId), batchId(batchId), streamId(streamId), contextId(contextId), indexId(indexId),
           deviceTaskType(deviceType), hostTaskType(hostType), modelId(modelId), connectionId(connectionId),
           startTime(startTime), endTime(endTime) {}
